defualt service

+++ src/main/java/com/dazzle/asklepios/client/setup/dto/ServiceSetupDTO.java
@@ -3,6 +3,7 @@ package com.dazzle.asklepios.client.setup.dto;
 import com.dazzle.asklepios.domain.enumeration.Currency;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 public record ServiceSetupDTO(
         Long id,
@@ -10,7 +11,7 @@ public record ServiceSetupDTO(
         String abbreviation,
         String code,
         String category,
-        Long price,
+        BigDecimal price,
         Currency currency,
         Boolean isActive,
         Long facilityId,
